% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn.r
\name{as_lookup}
\alias{as_lookup}
\title{Lookup autonomous system information for a vector of string IPv4 addresses}
\usage{
as_lookup(ips)
}
\arguments{
\item{ips}{character vector of IPv4 addresses}
}
\value{
A \code{data.frame} including \code{ips}, the numeric version of
        \code{ips}, the \code{as} (without the "\code{AS}" prefix and the
        \code{org}.
}
\description{
The MaxMind GeoLite autonomous system organization database is automatically
downloaded if not installed. You should update the database (roughly) twice
a month (which is roughly how often MaxMind updates the legacy product) using
the \code{link{update_as_db()}} function.
}
\note{
This product includes GeoLite data created by MaxMind, available from
      \href{http://www.maxmind.com}{MaxMind}.
}
\examples{
ips <- c("188.143.232.26", "193.189.117.147", "188.143.232.41",
         "167.114.224.11", "87.69.225.131", "14.106.226.74",
         "14.106.229.236", "1.227.120.42", "81.213.203.232")
as_info <- as_lookup(ips)
}

