% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cidr.r
\name{asn_table_to_trie}
\alias{asn_table_to_trie}
\title{Convert a \emph{pyasn} generated CIDR data file to a trie}
\usage{
asn_table_to_trie(asn_table_file)
}
\arguments{
\item{asn_table_file}{filename of dat file (can be gzip'd)}
}
\description{
Convert a \emph{pyasn} generated CIDR data file to a trie
}
\examples{
asn_table_to_trie(system.file("test", "rib.tst.gz", package="iptools"))
}
