% Generated by roxygen2 (4.0.1): do not edit by hand
\name{asnip}
\alias{asnip}
\title{Return a data frame of IPv4 to ASN & org mappings}
\usage{
asnip(ip, includeAS = TRUE, showMessages = FALSE)
}
\arguments{
\item{ip}{character vector of IPv4 addresses to lookup}

\item{includeAS}{(bool) whether to include or not include the "AS" prefix in the \code{asnum} column (default TRUE)}

\item{showMessages}{show/hide console messages (bool) default: do not show messages}
}
\value{
data frame of AS # & org information for the IP addresses
}
\description{
Uses the maxmind \code{GeoIPASNum.dat} binary file to perform an AS number & org
identification for a given IPv4 address and returns a data frame of results.
You should call \code{maxmindinit()} before calling \code{asnip()}, but \code{asnip}
will make a last-ditch effort to load the ASNum database from the default location
in the event you forget.
}
\details{
Values returned in the data frame:
\itemize{
  \item \code{ip}. original IP address (chr)
  \item \code{asn}. ASN (chr)
  \item \code{org}. Assigned org (chr)
}
}
\note{
vectorized
}
\examples{
\dontrun{
maxmindinit()
set.seed(1000000); asnip(randomIPs(10))
##                 ip     asn                         org
## 1       70.5.34.39  AS3651                      Sprint
## 2    79.32.183.102  AS3269       Telecom Italia S.p.a.
## 3     70.166.53.78 AS36801   New Wave Industries, Inc.
## 4  131.199.143.169  AS1341 Allen-Bradley Company, Inc.
## 5    160.150.52.98  AS1515        Headquarters, USAISC
## 6   242.119.216.60    <NA>                        <NA>
## 7     53.55.24.145 AS31399   Daimler Autonomous System
## 8   151.141.222.29 AS19956          BellSouth.net Inc.
## 9  243.118.250.204    <NA>                        <NA>
## 10   83.173.28.251 AS31441  Gagnaveita Reykjavikur ehf
}
}

