% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cidr_range}
\alias{cidr_range}
\title{IPv4 CIDR to long integer range}
\usage{
cidr_range(cidr)
}
\arguments{
\item{cidr}{IPv4 CIDR (str) dotted-decimal-slash-integer}
}
\value{
vector containing the minimum and maximum range integer values or \code{c(NA,NA)}
}
\description{
Converts IPv4 CIDR (e.g. "192.168.1.0/24") to vector containing the minimum and maximum range integer values
}
\examples{
\dontrun{
long2ip(cidr_range("192.168.1.0/24"))
## [1] "192.168.1.0"   "192.168.1.255"

long2ip(cidr_range("8.0.0.0/8"))
## [1] "8.0.0.0"       "8.255.255.255"

cr <- cidr_range("172.18.0.0/28")
sapply(cr[1]:cr[2], long2ip)
##  [1] "172.18.0.0"  "172.18.0.1"  "172.18.0.2"  "172.18.0.3"
##  [5] "172.18.0.4"  "172.18.0.5"  "172.18.0.6"  "172.18.0.7"
##  [9] "172.18.0.8"  "172.18.0.9"  "172.18.0.10" "172.18.0.11"
## [13] "172.18.0.12" "172.18.0.13" "172.18.0.14" "172.18.0.15"
}
}

