% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{classify_ip}
\alias{classify_ip}
\title{Identify whether an IP address is IPv4 or IPv6}
\usage{
classify_ip(ip_addresses)
}
\arguments{
\item{ip_addresses}{a vector of IPv4 or IPv6 IP addresses.}
}
\value{
a vector containing the class of each input IP address; either
"IPv4", "IPv6" or, for IP addresses that were not valid, "Invalid".
}
\description{
Identify the form (IPv4 or IPv6) of a vector
of IP addresses. This can also be used to validate IPs.
}

