% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expand_ipv6}
\alias{expand_ipv6}
\title{Expand an IPv6 address from an abbreviated version}
\usage{
expand_ipv6(ip_addresses)
}
\arguments{
\item{ip_addresses}{a vector of IPv6 IP addresses.}
}
\value{
a character vector of expanded IPv6 addresses
}
\description{
Expand an IPv6 address from an abbreviated version
}
\references{
\url{http://www.ipv6tf.org/index.php?page=meet/faqs&faq_id=1000&q=11}
}
