% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{geofile}
\alias{geofile}
\title{Initializes the maxmind library and opens the \code{GeoLiteCity.dat} file}
\usage{
geofile(datafile = "/usr/local/share/GeoIP/GeoLiteCity.dat")
}
\arguments{
\item{datafile}{full path (including filename) to the \code{GeoLiteCity.dat} file}
}
\description{
This function must be called before performing a lookup with \code{geoip()}.
The default full path spec defaults to \code{/usr/local/share/GeoIP/GeoLiteCity.dat}
and can be overriden by changing the value of the \code{datafile} parameter.
}
\details{
NOTE: You must manually retrieve the \code{GeoLiteCity.dat.gz} file from maxmind
via \url{http://geolite.maxmind.com/download/geoip/database/GeoLiteCity.dat.gz} and
unzip it to the location you will be specifying in \code{datafile}.

Any errors reading the file will result in a console message.
}
\examples{
\dontrun{
geofile()
}
}

