% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{geoip}
\alias{geoip}
\title{Return a data frame of geolcation values for a given IP address}
\usage{
geoip(ip)
}
\arguments{
\item{ip}{character string of IPv4 address to lookup}
}
\value{
data frame row of geolocation information
}
\description{
Uses the maxmind \code{GeoIPCity.dat} binary file to perform a geolocation for
a given IPv4 address and returns a data frame of geolocation records.
You must call \code{geofile()} before calling \code{geoip()}
}
\details{
Values returned in the data frame:
\itemize{
  \item \code{country.code}. ISO2 country code (chr)
  \item \code{country.code3}. ISO3 country code (chr)
  \item \code{region}. Abbreviated region name (chr)
  \item \code{region.name}. Full region name (chr)
  \item \code{city}. City name (chr)
  \item \code{postal.code}. Postal code (chr)
  \item \code{latitude}. Latitude (numeric)
  \item \code{lontigude}. Longitude (numeric)
  \item \code{time.zone}. Time zone (chr)
  \item \code{metro.code}. Metro code (int)
  \item \code{area.code}. Area code (int)
}
}
\note{
not vectorized
}
\examples{
\dontrun{
geofile()
geoip("20.30.40.50")
##  country.code country.code3  country.name region region.name         city
## 1           US           USA United States     VA    Virginia Falls Church
##   postal.code latitude longitude        time.zone metro.code area.code
## 1       22042   38.864  -77.1922 America/New_York        511       703
}
}

