% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_ranges.R
\name{get_all_country_ranges}
\alias{get_all_country_ranges}
\title{Fetch all country CIDR blocks}
\usage{
get_all_country_ranges()
}
\value{
named list of character vectors of CIDR blocks
}
\description{
Iterates through all the country codes and returns
a named list of CIDR blocks for all those countries.
}
\note{
This is an expensive operation as it pulls 249 files from
      \url{http://www.iwik.org/ipcountry/}. Try not to do this too often.
}
