% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gethostbyaddr}
\alias{gethostbyaddr}
\title{Returns all \code{PTR} records associated with an IPv4 address}
\usage{
gethostbyaddr(ipv4)
}
\arguments{
\item{ipv4}{character string (dotted-decimal) IPv4 address}
}
\value{
vector of hostnames (will usually be just a single host)
}
\description{
Takes in an IPv4 address and returns all IP reverse DNS entries for it. Not vectorized.
}
\examples{
\dontrun{
# reverse
gethostbyaddr(gethostbyname("dds.ec"))
## [1] "162.243.111.4"
}
}

