% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gethostbyname}
\alias{gethostbyname}
\title{Returns all \code{A} records associated with a hostname}
\usage{
gethostbyname(hostname)
}
\arguments{
\item{hostname}{character string of host to lookup}
}
\value{
vector IP addresses
}
\description{
Takes in a hostname and returns all IP DNS entries for it. Not vectorized.
}
\examples{
\dontrun{
# forward
gethostbyname("dds.ec")
## [1] "162.243.111.4"

# multiple return values
gethostbyname("google.com")
##  [1] "2607:f8b0:4006:806::100e" "74.125.226.14"
##  [3] "74.125.226.8"             "74.125.226.3"
##  [5] "74.125.226.6"             "74.125.226.4"
##  [7] "74.125.226.9"             "74.125.226.0"
##  [9] "74.125.226.2"             "74.125.226.1"
## [11] "74.125.226.5"             "74.125.226.7"
}
}

