% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gethostsbyaddr}
\alias{gethostsbyaddr}
\title{Vectorized version of \code{gethostbyaddr}}
\usage{
gethostsbyaddr(ipv4)
}
\arguments{
\item{ipv4}{vector of IPv4 address (character - dotted-decimal)}
}
\value{
list
}
\description{
Returns a list (slots are named by the input IPv4 addresses) with lookup results per slot
}
\examples{
set.seed(1492)
gethostsbyaddr(randomIPs(5))
}

