% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gethostsbyname}
\alias{gethostsbyname}
\title{Vectorized version of \code{gethostbyname}}
\usage{
gethostsbyname(hostname)
}
\arguments{
\item{hostname}{vector of host names (character - FQDN)}
}
\value{
list
}
\description{
Returns a list (slots are named by the input host names) with lookup results per slot
}
\examples{
gethostsbyname(c("google.com", "stackoverflow.com", "twitter.com", "r-bloggers.com"))
}

