% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hilbert_encode}
\alias{hilbert_encode}
\title{Encode an IPv4 address to Hilbert space}
\usage{
hilbert_encode(x, bpp = 8L)
}
\arguments{
\item{x}{IPv4 address}

\item{bpp}{Hilbert depth (max 12)}
}
\description{
Encode an IPv4 address to Hilbert space
}
