% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cidr.r
\name{host_count}
\alias{host_count}
\title{Return the number of hosts in a CIDR block}
\usage{
host_count(cidrs)
}
\arguments{
\item{cidrs}{character vector of IPv4 CIDRs}
}
\description{
TODO: ipv4 validation
}
\examples{
host_count("1.52.0.0/14")
}
