% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{data}
\name{ianaports}
\alias{ianaports}
\title{IANA Service Name and Transport Protocol Port Number Registry}
\format{A data frame with 13,527 rows and 12 variables}
\usage{
data(ianaports)
}
\description{
IANA Service Name and Transport Protocol Port Number Registry
\itemize{
  \item \code{Service.Name}. service name for the port assignment (chr)
  \item \code{Port.Number}. port(s) assigned (chr) - can be a range (e.g. 6629-6630)
  \item \code{Transport.Protocol}. transport protocol(s) of the port assignment (chr) - [\code{dccp}|\code{sctp}|\code{tcp}|\code{udp}]
  \item \code{Description}. explanation of the port assignment (chr)
  \item \code{Assignee}. Name and email address of the party to whom the assignment is made (chr)
  \item \code{Contact}. Name and email address of the Contact person for the assignment (chr)
  \item \code{Registration.Date}. date the port was registered (chr)
  \item \code{Modification.Date}. date of any modification to record (chr)
  \item \code{Reference}. A description of (or a reference to a document describing) the protocol or application using this port (chr)
  \item \code{Known.Unauthorized.Uses}. A list of uses by applications or organizations who are not the Assignee (chr)
  \item \code{Assignment.Notes}. Indications of owner/name change, or any other assignment process issue. (chr)
}
}
\note{
Last updated 2014-08-07
}
\seealso{
\itemize{
  \item IANA - \url{http://www.iana.org/assignments/service-names-port-numbers/service-names-port-numbers.xhtml}
  \item RFC6335 - \url{http://tools.ietf.org/html/rfc6335},
}
}
\keyword{datasets}

