% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ianaprotocolnumbers}
\alias{ianaprotocolnumbers}
\title{IANA Protocol Numbers}
\format{A data frame with 148 rows and 4 variables}
\usage{
data(ianaprotocolnumbers)
}
\description{
IANA Protocol Numbers
\itemize{
  \item \code{Decimal}. protocol number (chr)
  \item \code{Keyword}. short name/acronym of the protocol name (chr)
  \item \code{Protocol}. expanded name of the protocol (chr)
  \item \code{IPv6.Extension.Header}. has an IPv4 extension header (chr)
}
}
\note{
Last updated 2014-08-10
}
\seealso{
\itemize{
  \item IANA - \url{http://www.iana.org/domains/root/db}
  \item RFC5237 - \url{http://www.iana.org/go/rfc5237}
  \item RFC7045 - \url{http://www.iana.org/go/rfc7045}
}
}
\keyword{datasets}

