% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ianarootzonetlds.R
\docType{data}
\name{ianarootzonetlds}
\alias{ianarootzonetlds}
\title{IANA Root Zone Database}
\format{A data frame with 679 rows and 3 variables}
\usage{
data(ianarootzonetlds)
}
\description{
IANA Root Zone Database
\itemize{
  \item \code{Domain}. top level domain name in dot-tld format (chr) (e.g. "\code{.active}")
  \item \code{Type}. type of TLD (chr) - one of [\code{country-code}|\code{generic}|\code{generic-restricted}|\code{infrastructure}|\code{sponsored}|\code{test}]
  \item \code{Sponsoring.Organisation}. name of the entity sponsoring the TLD (chr)
}
}
\note{
Last updated 2014-08-09
}
\seealso{
\itemize{
  \item IANA - \url{http://www.iana.org/domains/root/db}
}
}
\keyword{datasets}

