% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{ip2long}
\alias{ip2long}
\title{Character (dotted-decimal) IPv4 Address Conversion to long integer}
\usage{
ip2long(ip)
}
\arguments{
\item{ip}{input character vector of IPv4 addresses (dotted-decimal)}
}
\value{
vector of equivalent long integer IP addresses
}
\description{
Convert IP addresses in character (dotted-decimal) notation to long integers
}
\examples{
\dontrun{
ip2long("24.0.5.11")
ip2long(c("24.0.5.11", "211.3.77.96"))
}
}

