% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ip_in_any}
\alias{ip_in_any}
\title{check if IP address falls within any of the ranges specified}
\usage{
ip_in_any(ip_addresses, ranges)
}
\arguments{
\item{ip_addresses}{character vector of IP addresses}

\item{ranges}{character vector of CIDR reanges}
}
\value{
a logical vector of whether a given IP was in any of the ranges
}
\description{
\code{ip_in_any} checks whether a vector of IP addresses
fall within any of the speficied ranges.
}

