% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ip_numeric_to_binary_string}
\alias{ip_numeric_to_binary_string}
\title{Convert a charcter vector of IPv4 addresses to a character vector of
bit strings.}
\usage{
ip_numeric_to_binary_string(input)
}
\arguments{
\item{input}{numeric vector of IP addresses}
}
\description{
Convert a charcter vector of IPv4 addresses to a character vector of
bit strings.
}
