% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{ip_random}
\alias{ip_random}
\title{generate random IPv4 IP addresses}
\usage{
ip_random(n)
}
\arguments{
\item{n}{the number of IP addresses to randomly generate.}
}
\value{
a character vector of randomly-generated IPv4 addresses,
in their dotted-decimal form.
}
\description{
\code{ip_random} generates random IP addresses.
These currently only follow IPv4 standards, since IPv6 addresses
are too large to be stored in R in their numeric form. All
IPs generated this way are valid.
}
\examples{
ip_random(1)
#[1] "49.20.57.31"
}
\seealso{
\code{\link{ip_to_numeric}} for converting \code{random_ips}'
output to its numeric form, and \code{\link{range_generate}} for
generating all IP addresses within a specific range.
}
