% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cidr.r
\name{ip_to_asn}
\alias{ip_to_asn}
\title{Match IP addresses to autonomous systems}
\usage{
ip_to_asn(cidr_trie, ip)
}
\arguments{
\item{cidr_trie}{trie created with \code{asn_table_to_trie()}}

\item{ip}{character vector or numeric vector of IPv4 addresses}
}
\description{
Match IP addresses to autonomous systems
}
\examples{
tbl <- asn_table_to_trie(system.file("test", "rib.tst.gz", package="iptools"))
ip_to_asn(tbl, "5.192.0.1")
}
