% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ip_to_binary_string}
\alias{ip_to_binary_string}
\title{Convert a numeric vector of IPv4 addresses to a character vector of
bit strings.}
\usage{
ip_to_binary_string(input)
}
\arguments{
\item{input}{character vector of IP addresses}
}
\description{
Convert a numeric vector of IPv4 addresses to a character vector of
bit strings.
}
