% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ip_to_numeric}
\alias{ip_to_numeric}
\title{convert a dotted-decimal IPv4 address to its numeric form.}
\usage{
ip_to_numeric(ip_addresses)
}
\arguments{
\item{ip_addresses}{a vector of IP addresses.}
}
\value{
a vector containing the numeric representation of \code{ip_addresses}.
If an IP is invalid (either because it's an Ipv6 address, or isn't an IP address
at all) the returned value for that IP will be -1.
}
\description{
\code{ip_to_numeric} takes IP addresses stored
in their human-readable representation ("192.168.0.1")
and converts it to a numeric representation (3232235521). Due to
limitations in the underlying software, and R's support for colossally
big numbers, this currently only works for IPv4 IP addresses.
}
\examples{
#Convert your local, internal IP to its numeric representation.
ip_to_numeric("192.168.0.1")
#[1] 3232235521
}

