% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cidr.r
\name{ips_in_cidrs}
\alias{ips_in_cidrs}
\title{Determine if a vector if IPv4 addresses are in a vector of CIDRs}
\usage{
ips_in_cidrs(ips, cidrs)
}
\arguments{
\item{ips}{character vector or numeric vector of IPv4 addresses}

\item{cidrs}{character vector or numeric vector of IPv4 CIDRs}
}
\value{
\code{data_frame} with \code{ips} column and a logical \code{in_cdir} column
}
\description{
Determine if a vector if IPv4 addresses are in a vector of CIDRs
}

