% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{ipv4heatmap}
\alias{ipv4heatmap}
\title{Create an IPv4 Hilbert curve heatmap}
\usage{
ipv4heatmap(ips, colors = NA, cb_pal = "PuOr", alpha = FALSE,
  legend = FALSE)
}
\arguments{
\item{ips}{character vector of IPv4 addresses}

\item{colors}{character vector (5 elements) of colors to be used in the plot.
Each color maps to the number of IP addresses in the netblock (~log scale).
By default, it will use ColorBrewer "PuOr" range}

\item{cb_pal}{named RColorBrewer palette to use (using \code{colors} overrides
any value used here)}

\item{alpha}{scale pixel alpha along with color}

\item{legend}{if you want the legend, then set this to TRUE.}
}
\value{
list containing a ggplot2 object (\code{gg}) and the x,y point data
        (\code{dt}) - which is a \code{data.table}).
}
\description{
Return a minimally annotated \code{ggplot2} object of a Hilbert-space heatmap for
a set of IPv4 addresses.
}
\examples{
ips <- c("37.113.19.46", "3.188.210.86", "81.97.190.213", "36.242.75.255",
"37.85.209.22", "79.246.149.132", "124.43.169.59", "15.114.34.91",
"99.202.51.14", "97.211.252.39", "78.43.125.169", "28.245.39.90",
"12.80.98.188", "1.139.230.141", "113.133.95.189", "98.194.96.20",
"109.57.232.176", "47.49.35.6", "14.187.171.0", "62.186.25.75")
hm <- ipv4heatmap(ips)

## plot it
# hm$gg

## inspect the points data.table:
str(hm$dt)
}

