% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ipv6_to_bytes}
\alias{ipv6_to_bytes}
\title{Convert a character vector of IPv6 addresses to a list of raw vectors of bytes}
\usage{
ipv6_to_bytes(input)
}
\arguments{
\item{input}{input IPv6 string vector}
}
\description{
Convert a character vector of IPv6 addresses to a list of raw vectors of bytes
}
\examples{
c("2001:db8::1",
  "2001:e42:101:2:202:181:99:52",
  "2400:8500:1801:417:118:27:35:213",
  "x",
  "2a02:2770:8:0:21a:4aff:fe96:7a47",
  "2400:2413:32c0:8:21a:92ff:fe22:c7b3",
  "2001:44b8:3138:c570:250:56ff:fe9c:c19b",
  "240f:a2:2e5:1:214:c2ff:fec8:1673",
  "2001:e42:102:1103:153:121:36:109",
  "2401:2500:203:2f:153:127:108:158"
) -> tst6

ipv6_to_bytes(tst6)
}
