% Generated by roxygen2 (4.0.1): do not edit by hand
\name{maxmindinit}
\alias{maxmindinit}
\title{Initializes the MaxMind library and opens the \code{GeoLiteCity.dat} & \code{GeoLiteASNum.dat} file}
\usage{
maxmindinit(citydata = "/usr/local/share/GeoIP/GeoLiteCity.dat",
  asndata = "/usr/local/share/GeoIP/GeoIPASNum.dat")
}
\arguments{
\item{citydata}{full path (including filename) to the \code{GeoLiteCity.dat} file}

\item{asndata}{full path (including filename) to the \code{GeoLiteASNum.dat} file}
}
\description{
This function must be called before performing a lookup with \code{geoip()} or \code{asnip()}.
Each full path spec prefix defaults to \code{/usr/local/share/GeoIP/}
and can be overriden by changing the value of the \code{citydata}  & \code{asndata} parameters.
}
\details{
NOTE: You must manually retrieve the \code{GeoLiteCity.dat.gz} \code{GeoLiteASNum.dat.gz} files from MaxMind
via \url{http://geolite.maxmind.com/download/geoip/database/GeoLiteCity.dat.gz} &
\url{http://geolite.maxmind.com/download/geoip/database/GeoLiteASNum.dat.gz}  and
unzip each to the locations you will be specifying in the parameters.

Any errors reading the files will result in a console message.
}
\examples{
\dontrun{
maxmindinit()
}
}

