% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generators.R
\name{random_ips}
\alias{random_ips}
\title{generate random IPv4 IP addresses}
\usage{
random_ips(n)
}
\arguments{
\item{n}{the number of IP addresses to randomly generate.}
}
\value{
a character vector of randomly-generated IPv4 addresses,
in their dotted-decimal form.
}
\description{
\code{random_ips} generates random IP addresses.
These currently only follow IPv4 standards, since IPv6 addresses
are too large to be stored in R in their numeric form. All
IPs generated this way are valid.
}
\examples{
random_ips(1)
#[1] "49.20.57.31"
}
\seealso{
\code{\link{ip_to_numeric}} for converting \code{random_ips}'
output to its numeric form.
}

