% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iptools.R
\name{randomIPs}
\alias{randomIPs}
\title{Generate random sequence of IPv4 addresses}
\usage{
randomIPs(size)
}
\arguments{
\item{size}{number of IP addreses to generate}
}
\value{
character vector
}
\description{
Returns a character vector of randomly generated (but valid) IPv4 addresses
}
\examples{
set.seed(1492)
randomIPs(10)
}

