% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{range_boundaries}
\alias{range_boundaries}
\title{calculate the maximum and minimum IPs in an IP range}
\usage{
range_boundaries(ranges)
}
\arguments{
\item{ranges}{a vector of IP ranges. Currently only IPv4 ranges
work.}
}
\value{
a list of character vectors, each one consisting of the maximum
and minimum IPs within the equivalent range.
}
\description{
when provided with a vector of IP ranges
("172.18.0.0/28"), \code{range_boundaries} calculates the
maximum and minimum IP addresses in that range.
}
\examples{
range_boundaries("172.18.0.0/28")
# [[1]]
# [1] "172.18.0.0"  "172.18.0.15"
}
\seealso{
\code{\link{ip_in_range}} to calculate if an IP address
falls within a particular range, or \code{\link{ip_to_numeric}} to
convert the dotted-decimal notation of returned IP addresses to their
numeric representation.
}

