% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn.r
\name{update_as_db}
\alias{update_as_db}
\title{Update the autonomous system database}
\usage{
update_as_db()
}
\description{
Use this function to update the autonomous system database.
The legacy MaxMind databases update (roughly) twice a month
}
\note{
This product includes GeoLite data created by MaxMind, available from
      \href{http://www.maxmind.com}{MaxMind}.
}

