% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{v6_scope}
\alias{v6_scope}
\title{Return the scope of an IPv6 address (string)}
\usage{
v6_scope(ip_addresses)
}
\arguments{
\item{ip_addresses}{a vector of IPv6 IP addresses.}
}
\value{
a numeric vector of scopes
}
\description{
Return the scope of an IPv6 address (string)
}
\references{
\url{https://tools.ietf.org/html/rfc4007}
}
