% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iptools.R
\name{validateCIDR}
\alias{validateCIDR}
\title{Validate IPv4 CIDRs in dotted-decimal slash notation}
\usage{
validateCIDR(cidr)
}
\arguments{
\item{cidr}{vector of IPv4 CIDRs (character - dotted-decimal slash notation)}
}
\value{
named logical vector
}
\description{
Input CIDRs are considered valid if they consist of a valid IPv4 address in dotted-quad format followed by a forward slash (/) and a bit mask length (1-32)
}
\examples{
validateIP("127.0.0.1/32")
validateIP("127.0/8")
validateIP(c("127.0.0.256/32", "127.0.0.0", "127.0.0.1/33", "127.0.0.1/32"))
}

