% Generated by roxygen2 (4.0.1): do not edit by hand
\name{validateIP}
\alias{validateIP}
\title{Validate IPv4 addresses in dotted-decimal notation}
\usage{
validateIP(ip)
}
\arguments{
\item{ip}{vector of IPv4 address (character - dotted-decimal)}
}
\value{
named logical vector
}
\description{
Input addresses are considered valid if they consist of 1-4 octets separated by periods
}
\examples{
validateIP("127.0.0.1")
validateIP("127.0")
validateIP(c("127.0.0.1", "255.255.255.255", "500.1.1.1", "30.2.2"))
}

