% Generated by roxygen2 (4.0.1): do not edit by hand
\name{geoip}
\alias{geoip}
\title{Return a data frame of geolcation values for IPv4 address}
\usage{
geoip(ip, showMessages = FALSE)
}
\arguments{
\item{ip}{character vector of IPv4 addresses to lookup}

\item{showMessages}{show/hide console messages (bool) default: do not show messages}
}
\value{
data frame of geolocation information for the IP addresses
}
\description{
Uses the maxmind \code{GeoIPCity.dat} binary file to perform a geolocation for
a given IPv4 address and returns a data frame of geolocation records.
You must call \code{geofile()} before calling \code{geoip()}
}
\details{
Values returned in the data frame:
\itemize{
  \item \code{ip}. original IP address (chr)
  \item \code{country.code}. ISO2 country code (chr)
  \item \code{country.code3}. ISO3 country code (chr)
  \item \code{region}. Abbreviated region name (chr)
  \item \code{region.name}. Full region name (chr)
  \item \code{city}. City name (chr)
  \item \code{postal.code}. Postal code (chr)
  \item \code{latitude}. Latitude (numeric)
  \item \code{lontigude}. Longitude (numeric)
  \item \code{time.zone}. Time zone (chr)
  \item \code{metro.code}. Metro code (int)
  \item \code{area.code}. Area code (int)
}
}
\note{
vectorized
}
\examples{
\dontrun{
geofile()
geoip(c("24.24.24.24", "42.42.42.42", "8.8.8.8"))
##            ip country.code country.code3       country.name region region.name      city
## 1 24.24.24.24           US           USA      United States     NY    New York Deer Park
## 2 42.42.42.42           KR           KOR Korea, Republic of     NA          NA        NA
## 3     8.8.8.8           US           USA      United States     NA          NA        NA
##   postal.code latitude longitude        time.zone metro.code area.code
## 1       11729  40.7627  -73.3227 America/New_York        501       631
## 2          NA  37.5700  126.9800       Asia/Seoul          0         0
## 3          NA  38.0000  -97.0000               NA          0         0
}
}

