% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iptools.R
\docType{package}
\name{iptools}
\alias{iptools}
\alias{iptools-package}
\title{A package to quickly and easily handle IP addresses.}
\description{
This package provides functions for handling IP addresses,
including (but not limited to) validating, generating and resolving
IP addresses, as well as converting them between different formats.
}

