% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptools-package.R
\docType{package}
\name{iptools}
\alias{iptools}
\alias{iptools-package}
\title{A package to quickly and easily handle IP addresses.}
\description{
A toolkit for manipulating, validating and testing 'IP' addresses and
 ranges, along with datasets relating to 'IP' addresses. Tools are also provided
 to map IPv4 blocks to country codes. While it primarily has support for the 'IPv4'
 address space, more extensive 'IPv6' support is intended.
}
